+++ .ansible/collections/ansible_collections/netapp/ontap/plugins/modules/na_ontap_file_security_permissions_acl.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2022-2025, NetApp, Inc
+# (c) 2022-2026, NetApp, Inc
 # GNU General Public License v3.0+  (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from __future__ import absolute_import, division, print_function
@@ -195,10 +195,33 @@ options:
     type: str
     default: error
 
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.6.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
 notes:
   - Supports check_mode.
   - Only supported with REST and requires ONTAP 9.9.1 or later.
   - SLAG requires ONTAP 9.10.1 or later.
+  - Supports AWS Lambda proxy functionality. See README for example usage.
 '''
 
 EXAMPLES = """
@@ -296,8 +319,12 @@ class NetAppOntapFileSecurityPermissionsACL:
             acl_user=dict(required=True, type='str'),
             validate_changes=dict(required=False, type='str', choices=['ignore', 'warn', 'error'], default='error'),
         ))
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
+            required_if=[
+                ('use_lambda', True, ['lambda_config']),
+            ],
             supports_check_mode=True
         )
         self.svm_uuid = None